feat:函数式调用组件

+++ src/views/cart/index.vue
@@ -111,6 +111,7 @@
 import GoodRelevant from '@/views/goods/components/goods-relevant'
 import { useStore } from 'vuex'
 import Message from '@/components/library/Message'
+import Confirm from '@/components/library/Confirm'
 import CartNone from './components/cart-none.vue'
 export default {
   name: 'XtxCartPage',
@@ -127,9 +128,15 @@ export default {
     }
     // 删除
     const deleteCart = skuId => {
-      store.dispatch('cart/deleteCart', skuId).then(() => {
-        Message({ type: 'success', text: '删除成功' })
-      })
+      Confirm({ text: '亲，您是否确认删除商品' })
+        .then(() => {
+          store.dispatch('cart/deleteCart', skuId).then(() => {
+            Message({ type: 'success', text: '删除成功' })
+          })
+        })
+        .catch(e => {
+          console.log('取消')
+        })
     }
     return { checkOne, checkAll, deleteCart }
   }
